Add BLS Aggregation window waiting after quorum (#152)
Add BLS aggregation window-waiting-after-quorum feature. Once the
aggregator reaches quorum, instead of sending the aggregated signature
response, it will wait for a `window_duration` time. During that window
time, it will continue to receive and aggregate new signatures, allowing
to reduce the gas cost of validating the aggregation on chain. When the
window time finishes, the aggregation is sent.
In the case that task's expiry time is reached during that window, the
aggregation response is sent.

To implement this feature, the aggregation loop was modified. There are
3 posible cases:
- receiving a new signature,
- expiration timer,
- window timer (implemented using a channel and a tokio task).

Each time a new signature is received, it checks if quorum is reached.
Once it's reached, the window timer fires. The loop continues until one
of the two timers finish (window timer or task expiration).

This tests cases were added:
- signatures are processed during window after quorum
- if quorum has been reached and the task expires during window, the
response is sent
- if window duration is zero, no signatures are aggregated after
reaching quorum
- no signatures are aggregated after window

---------

Co-authored-by: Pablo Deymonnaz <deymonnaz@gmail.com>