ci: add CI job to check whether anvil state is up to date (#237)
Fixes #234

This PR introduces the following changes
* Adds a CI job to check if the anvil dump state is up to date.
* Fixes the anvil's initial timestamp, otherwise the dump state will be
different each time (since the timestamps are included in the state).
* Updates the anvil dump state.

The CI job works by generating a new anvil state and comparing it
against the existing one using the `diff` command.
If the new anvil state is different it means the existing one is
outdated and therefore the CI will fail.
**Note:** before performing the comparison between the existing and the
new state, both files are sorted since the JSON format does preserve
item order.

- [ ] New features are tested and documented
- [ ] PR has one of the `changelog-X` labels (if applies)
- [ ] Code deprecates any old functionality before removing it