ci: fix job that checks anvil state is up-to-date (#252)
### Motivation
The previous CI job that checked whether the anvil state was up to date
was actually not performing any comparison at all (it was mistakenly
comparing empty files).

Checking that the anvil state is up to date by re-generating it and
comparing it with the existing one was not very practical since this
approach is very time-sensitive: any difference in the timestamps
generates a different state, which the CI wrongly takes as an outdated
state.

### What Changed?
This PR introduces the following changes in the CI:
* Removes the step that compares the anvil state in the tests CI job,
and simply checks that the script to generate the state works.
* Adds a new job that checks whether the anvil state is updated (it runs
only when the contracts change).


### Reviewer Checklist

- [ ] New features are tested and documented
- [ ] PR updates the changelog with a description of changes
- [ ] PR has one of the `changelog-X` labels (if applies)
- [ ] Code deprecates any old functionality before removing it

---------

Co-authored-by: Tomás Grüner <47506558+MegaRedHand@users.noreply.github.com>