feat: add update socket function (#268)
### What Changed?
Adds update socket function, that was implemented in Go SDK but not in
this. Also adds a simple test case.

Note: Since changelog is not implemented yet in `dev` branch, PR #251
should be merged first, and then there should be created a PR that
copies the changelog file from main branch to dev.

### Reviewer Checklist

- [ ] New features are tested and documented
- [ ] PR updates the changelog with a description of changes
- [ ] PR has one of the `changelog-X` labels (if applies)
- [ ] Code deprecates any old functionality before removing it