docs: update release process doc (#456)
### What Changed?

This PR updates the release process doc to include instructions on how
to release through `release-plz`.
It also includes a new section about running the release workflow on
branches apart from `dev` or `main`.

As an additional change, it removes the `publish_no_verify` flag from
the `release-plz` config, which was specified in an attempt to fix the
error we were getting when publishing, but it's unrelated to the actual
fix so now it can be removed.

### Reviewer Checklist

- [ ] New features are tested and documented
- [ ] PR updates the changelog with a description of changes
- [ ] PR has one of the `changelog-X` labels (if applies)
- [ ] Code deprecates any old functionality before removing it