feat(v2): add operator example (#436)
### What Changed?

This PR adds an example of how to initialize and start a registered
`operator`. The idea is to show what the dev experience would be to have
an operator begin listening for tasks and processing them.

Since I started from this
[PR](https://github.com/Layr-Labs/eigensdk-rs/pull/435), the diff
appears quite large + the bindings. The example is located in
`examples/operator`


### Reviewer Checklist

- [ ] New features are tested and documented
- [ ] PR updates the changelog with a description of changes
- [ ] PR has one of the `changelog-X` labels (if applies)
- [ ] Code deprecates any old functionality before removing it

---------

Co-authored-by: Pablo Deymonnaz <pdeymon@fi.uba.ar>
Co-authored-by: Tomás Grüner <47506558+MegaRedHand@users.noreply.github.com>