feat(v2): handle edge cases in operator registration  (#555)
### What Changed?
This PR adds a validation step when creating an operator to handle an
`OperatorRegistrationConfig`.

If the user provides an `OperatorRegistrationConfig` struct, the
operator creation process will:
1. Register the operator globally with EigenLayer
2. Deposit ERC-20 tokens into the specified strategies
3. Configure allocation magnitudes for those strategie
4. Register for the specified operator sets in each AVS

The goal is to provide a declarative registration system: users declare
the desired end state for example, "operator registered in an AVS, with
stakes x, y, and z in strategies A, B, and C" and the SDK automatically
executes all necessary steps to reach that state.

---

TODO:
- Add tests
- Check how to handle possible errors, we want to panic or log an error
and continue with the registration?

### Reviewer Checklist

- [ ] New features are tested and documented
- [ ] PR updates the changelog with a description of changes
- [ ] PR has one of the `changelog-X` labels (if applies)
- [ ] Code deprecates any old functionality before removing it